reorganize as a python package
