+++ create_executable_linux.sh
@@ -2,4 +2,4 @@
 
 source ./.venv/bin/activate
 echo 'Current python env ~~~>' $(which python3)
-python3 -m nuitka --standalone --plugin-enable=pyqt5 --enable-console h5temperature.py
+python3 -m nuitka --standalone --plugin-enable=pyqt5 run.py
