+++ create_executable_windows.bat
@@ -1,5 +1,6 @@
 python -m venv .venv 
 call .\.venv\Scripts\activate.bat
 python -m pip install -r requirements.txt
-python -m nuitka --standalone --follow-imports --enable-console --enable-plugin=pyqt5 h5temperature.py
+python -m pip install nuitka
+python -m nuitka --standalone --enable-plugin=pyqt5 run.py
 PAUSE
