+++ run.py
@@ -1,0 +1,3 @@
+from h5temperature import main
+
+main()
